Fix most issues caught by eslint.
Use eslint-plugin-react.

Ignore the generated files with lint.

Ignore an error in lint that is not correct.

Mostly cleaning up imports and spaces.  Fixes several places where the
wrong variable was used (real bugs) caught by the linter.

There are still a few lint issues left but this takes care of all the
ones that are safe to change.

Partially deals with #148