Add debug mode to production.
When in production (packaged mode) add --debug option to show dcrd and
dcrwallet output and info to stdout.  Otherwise there will be no
output to stdout.

The logs data is available in the app specific log files even without
the debug flag.

The behavior should be the same on OSX when started from a terminal
and Linux.

Closes #138