Give user some feedback for daemon syncing (#157)
Due to various constraints on paths of communication, we currently 
can't easily determine what the progress on a daemon/wallet block 
syncness is.

To give users some sort of feedback, we've decided to calculate a 
rough block height that a daemon "should be at" given the current 
date. So, for example, on testnet: It's been roughly 350 days since 
the chain started. With block time set in the dcrd params we can 
roughly expect 720 blocks mined each day on testnet. Then we 
multiple by 95% to accomodate for a bit of variance. So the current 
"neededBlocks" value is 239864 which is just under our current 
height of ~242000.

Obviously this is not ideal, but should be a good replacement in the 
meantime until we have another option.