Add logic for mined and unmined transactions (#192)
Previously, only mined transactions were handled from the getTransactions 
rpc and they were all appended to the transactions state array. Now unmined 
and mined are dealt with in a relatively smart way (checking if they already 
exist and don't add if so). This should set up decrediton to allow for pagination
 in a coming PR as well.