Add checks, options, and docs to docker builds.
This makes the docker build run the linter and fails if the linter
fails.  This will also catch compile errors.  It is necessary for
using this in a CI system, but also generally useful.

Added the ability to specify a github id to build from (to build a
fork).

Add some doc in the README.md for the docker build.