[politeia] Ignore listing after wallet closed
This changes the logic for fetching Politeia data such that, if the
wallet is detected as closed (or changed) then the results or any errors
are ignored.

This is necessary because the external access to Politeia can be slow
(due to server load) or the call to the wallets CommittedTickets grpc
endpoint may take some time to complete if the wallet has many tickets.

An user closing the wallet while the results are still being fetched
would cause an undersirable snackbar error message without this fix.