[mytickets] Filter duplicate non-wallet-controlled
This adds a filter to the GetTickets action, such that wallets that have
tickets and votes/revocations without the corresponding voting
authority, a duplicate error isn't thrown.

The source of this error is the wallet.GetTicket grpc endpoint not
considering tickets not from the wallet as voted, and incorrectly
sending their status as missed/expired.