[ln] Initial Version
This is the first version of an embedded LN wallet UI on decrediton.

Current features:
  - Startup, link dcrlnd to the currently running wallet and unlock it
  - Open/Close channels
  - Create invoices
  - Send payments
  - Enable autopilot
  - Show logs

The current implementation relies on the remote wallet functionality of dcrlnd.

For the moment, in order to enable support to LN, users need to manually edit
the global decrediton config file to set `"ln_enabled: true"`.