[ln] Add syncedToChain check at connect
This adds a check during node connection to ensure ln operations are only
performed after the node is synced to chain.

It also fixes a bug where you could go to another page, come back to the
ConnectPage view and try to connect again while a previous connection attempt
was still ongoing.