[balance] Fix commit balances (#2217)
Now dcrwallet tracks the lockedByTickets balance per-commit correctly
instead of only for owned tickets. So this commit fixes the charts
balance to account for that.

It collapses lockedByNonWallet and lockedByWallet into locked and
stops tracking immature tickets given the wallet doesn't differentiate
between locked+immature and locked+mature.

This downgrades the react-intl version to 3.3.1 due to custom messages
formatting being broken and therefore the stake reward % not being displayed
correctly.