[trezor] Fix send all funds
This fixes the send all funds. The send all feature uses a change address
to create the output so trezor was assuming and generating a wallet address
to send the funds to.

This fixes the problem by checking whether the change address really is from
the wallet and only encoding as a trezor-generated address if it is.