[trezor] RC tweaks (#2308)
* [trezor] Add trezor_debug back to global config

* [trezor] Fix sending transaction double attempt

This fixes trying to double send a transaction when using trezor.

The SendTransactionButton is responsible for determining whether the wallet
is using trezor or not and dispatches the correct sign attempt so there's
no need to try to send again.

* [trezor] Fix send all funds

This fixes the send all funds. The send all feature uses a change address
to create the output so trezor was assuming and generating a wallet address
to send the funds to.

This fixes the problem by checking whether the change address really is from
the wallet and only encoding as a trezor-generated address if it is.

* [shared] Fix VerticalAccordion children styles

The children (content) of the vertical accordion should not get the same
classnames as the accordion itself.

* [actions] Sync on SPV after creating watch only wallet

Previously creating a watching only wallet started an RPC sync which would
fail if the wallet was actually configured as SPV.

* [deps] Update trezor.js version

* [trezor] Use local protobuf definitions

This changes the configuration of the BridgeV2 transport to use local
definitions for the protobuf messages instead of remotely fetching them.

The current definitions that are online at wallet.trezor.io/config.bin are
missing some messages needed to recover a Model T wallet, therefore the
messages were copied from the upstream trezor/connect repo.

* [trezor] Add minimum firmware version check

This adds a check for the minimum version of hardware that correctly
supports decred so that users don't attempt to use an incompatible
wallet.