[ln] Add choice to create LN account during setup
This adds a toggle that lets users choose whether to create a new LN-specific
wallet account when first initializing an LN wallet.

This way users are incentived to not use their default wallet account which
is likely to have the majority of their funds, which could cause many channels
to be automatically opened if the autopilot feature was selected.