[inputs] Improve DcrInput usage
This changes the DcrInput component to require only a single prop as
input ('amount').

Having to track both 'value' and 'amount' as consumer of DcrInput
was cumbersome and led to some mistakes such as in the ticket buyer
input that specifies the balance to maintain.

This improves the component so that only a single set of props
(amount/onChangeAmount) need to be specified and the component behaves
as expected.