[trezor] Send sequence for prevtx inputs
The sequence field for inputs of prevtxs was not being sent to the device,
causing a failure to verify that the hash of the previous transaction and its
contents matched.

This error could be triggered if the trezor wallet received funds from a
non-dcrwallet sender that did not set the Sequence field of the authored
transaction to MaxTxInSequenceNum.