[launcher] Fix SPV choice triggering daemon launch
This fixes an instance where selecting to use SPV for the first time on a new
install would cause the dcrd full node to run.

The problem was that to speed up the blockchain download, the node was started
after SPV choice and before the wallet tutorial but it was started
unconditionally, without considering the previous choice to only run in SPV
mode.