[home] Fix active ticket count in SPV wallets
SPV wallets don't track the live and missed ticket count only the unspent.
Therefore we introduce a new selector to track the active ticket count as
intended for display purposes in the home page.

This correctly tracks the immature+unspent count in SPV wallets and
immature+live in RPC wallets.