[cspp] integrating cspp into decrediton (#2475)
This PR is a continuation of the integration of cspp at decrediton. #2455 

-  It makes `mixedAccounts` not being able to receive funds (not showing
 addresses at receive page);
- It makes only possible to send funds from `mixedAccount` at send page; 
- It adds a `isMixTx` method which will help us determine which tx is mixed 
or not, so we will be able to flag them

    This way we will be able to restore private wallets and guess which 
account is the mixed account. 

    This code is based on 
https://github.com/decred/dcrdata/blob/master/txhelpers/cspp.go (thanks @chappjc ).

I think it would be good to add the ability to send funds from other accounts 
in case the user want to (mark a checkbox, which notifies him about the 
lack of privacy or the possibility of reducing the anonimous set where 
these inputs may be spent). 

After this being merged, we will still need to:
 - restore private wallets;
 - Add tutorial for begginers;
 - styling