[LN] Improve connection procedure
This makes the connection procedure to the underlying dcrlnd node services more
reliable by trying again when an EUNIMPLEMENTED error is returned.

Due to the gRPC server being restarted after an unlock, depending on the exact
timings, we might still connect to the wallet unlocker service so trying to
use the general services would fail.