TrezorActions: Allow updating firmware v2. (#2932)
We are unable to update firmware from TrezorActions.js. This is believed
to be due to the size of the firmware and the limitations of this
electron child process. We send the firmware to the main process to push
to the device. The main process uses a temporary connection to trezor in
order to accomplish this.

Send path to ipc.

Better handle update firmware failures.

For some reason importing connect in main_dev/ipc.js was not working
when building for release. Moving connect fixes it...