[vspd] Add processing of Managed and Unmanaged tickets (#3121)
This diff adds processing of managed tickets and unmanaged (solo) tickets. Now on restores
first process managed it called and updates the status of any live ticket that is found across
all available VSPs. Then if any are still found but not managed, they are given the chance
to process unmanaged and get their fees paid for solo tickets.

Also checks are done on normal starts/restarts to make sure all live tickets are properly managed
and without error.