Fix declaration for pi-ui and regen yarn.lock.
This format matches the dependency declaration for another
Decred dependency, node-addon-loader.

Regenerating yarn.lock using this fix modifies the "resolved" line
for fetching the dependency's source, in a format that is understood
by the Nix package manager after converting the yarn.lock using
yarn2nix.  Previously, the HTML for the GitHub project was fetched
instead, which failed checksum matches and never would have been usable
in the build.

To avoid changing the version of pi-ui in the build, the package.json
references the Git sha1 of the previously resolved commit.  This can
be changed to a branch or version specification when it needs to be
updated later.