Introduce preload script
This introduces a preload script to the app. This script is generated via
a new webpack config, which watches over changes to regenerate it as needed
when running via `yarn dev`.

Currently the only API reintroduced via this preload script into the main
world is the wallet/fs module. In future commits, other APIs will be
compatibilized to the preload script and will be offered through it.