Disable nodeIntegration for UI code
Now that all code that relies on node APIs is either using a polyfill or
is shimmed through to the preload script layer, this disables nodeIntegration
for the main UI code.

It also changes the target of the webpack config that generates the UI code
to use the "web" target, which is needed to disable nodeIntegration.