Add gRPC client tracking
This adds a client tracking layer to gRPC clients returned by the wallet pkg.

Instead of directly returning the gRPC client instance, the wallet keeps track
of the client and returns a random string ID to the UI code. Subsequent calls
to gRPC functions fetch the actual client from the tracked map.

This is needed due to contextIsolation not allowing non-primitive values from
crossing the preload -> ipcRenderer layers.