libdexc: update for dex 0.3
This updates for the changes in the dcrdex/client APIs.

This also prepares the CoreAdapter for a language to be specified on
construction of Core and WebServer.

The init method of CoreAdapter is modified for the new InitializeClient
syntax, which includes a seed []byte input argument for restoring.
This can be nil to generate fresh account keys.  Given new "restore from
seed?" dialogs, DEX accounts may be restored from seed.

A new method, (*CoreAdapter).discoverAcct, is created for use when and
if the init call is updated with an optional restoration seed. This
would be used to improve UX when discovering existing paid DEX accounts
as described in the docs for discoverAcct.

app: set DEX locale