i18n: translator: wrap strings list in own scroller
This quick hack does the job to fix the annoying need to scroll back and
forth with long lists of strings.

It is not great, it would be better to stretch the left-hand strings
list to use all vertical space on any display. I'm not familiar with the
Grid layout too much to do it properly, but we can improve it later.

Also, I'm not sure about the purpose of the Deleted Entries section but
I suspect it could be moved as a 4th radio button in the top-left. If
that happens to be a good idea, the proper scrolling fix should be done
afterwards.