[LN] Show startup progress and log line (#3643)
This adds a linear progress component during the router pruning stage of the
LN wallet startup. It also adds the last log line of the dcrlnd process to the
startup UI and removes the 60 second limit of the startup sync time.

These help mitigate issues when opening an LN wallet that has been offline for
a long time.